+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog of lizard-graph
 ===================================================
 
 
-0.1.2 (unreleased)
-------------------
+0.2 (2011-12-08)
+----------------
 
 - Changed model GraphLayout to an abstract class GraphLayoutMixin. The
   fields are now directly in GraphItem.